requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.46.0
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==3.0.4
eventlet>=0.24.1
Flask-SocketIO~=5.4.0
pandas~=2.2.3
tiktoken~=0.7.0
scipy~=1.14.1
PyPDF2~=3.0.1
SQLAlchemy~=2.0.35
WTForms~=3.1.2
python-docx~=1.1.2
Markdown~=3.7
beautifulsoup4~=4.12.3
google-api-python-client~=2.149.0
google-auth-oauthlib~=1.2.1
protobuf~=5.28.3
requests~=2.32.3
numpy~=2.1.1
matplotlib~=3.9.2
cachetools~=5.5.0
seaborn~=0.13.2
pdfplumber~=0.11.4
openpyxl~=3.2.0b1
Flask-APScheduler~=1.13.1
pdf2image~=1.17.0
pillow~=11.0.0
django-scheduler~=0.10.1
schedule~=1.2.2
plaid-python~=28.0.0
urllib3~=2.2.3